pyproject.toml
[project]
name = "rag-sample"
version = "0.1.0"
description = "RAG (Retrieval-Augmented Generation) CLI Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "click>=8.3.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-ollama>=0.3.10",
    "ollama>=0.4.9",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]

[project.scripts]
rag = "src.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pillow>=10.0.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: ユニットテスト（外部依存なし）",
    "integration: 統合テスト（外部依存あり）",
    "slow: 実行時間が長いテスト",
]
